// src/lib/interfaces/rustplus.ts
export interface AppTime {
    /**
     * @generated from protobuf field: float dayLengthMinutes = 1
     */
    dayLengthMinutes: number;
    /**
     * @generated from protobuf field: float timeScale = 2
     */
    timeScale: number;
    /**
     * @generated from protobuf field: float sunrise = 3
     */
    sunrise: number;
    /**
     * @generated from protobuf field: float sunset = 4
     */
    sunset: number;
    /**
     * @generated from protobuf field: float time = 5
     */
    time: number;
}
// ...
class AppTime$Type extends MessageType<AppTime> {
    constructor() {
        super("rustplus.AppTime", [
            {
                no: 1,
                name: "dayLengthMinutes",
                kind: "scalar",
                T: 2 /*ScalarType.FLOAT*/
            },
            {
                no: 2,
                name: "timeScale",
                kind: "scalar",
                T: 2 /*ScalarType.FLOAT*/
            },
            {
                no: 3,
                name: "sunrise",
                kind: "scalar",
                T: 2 /*ScalarType.FLOAT*/
            },
            {
                no: 4,
                name: "sunset",
                kind: "scalar",
                T: 2 /*ScalarType.FLOAT*/
            },
            { no: 5, name: "time", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ }
        ]);
    }
    create(value?: PartialMessage<AppTime>): AppTime {
        const message = globalThis.Object.create(this.messagePrototype);
        message.dayLengthMinutes = 0;
        message.timeScale = 0;
        message.sunrise = 0;
        message.sunset = 0;
        message.time = 0;
        if (value !== undefined)
            reflectionMergePartial<AppTime>(this, message, value);
        return message;
    }
    internalBinaryRead(
        reader: IBinaryReader,
        length: number,
        options: BinaryReadOptions,
        target?: AppTime
    ): AppTime {
        const message = target ?? this.create(),
            end = reader.pos + length;
        while (reader.pos < end) {
            const [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* float dayLengthMinutes */ 1:
                    message.dayLengthMinutes = reader.float();
                    break;
                case /* float timeScale */ 2:
                    message.timeScale = reader.float();
                    break;
                case /* float sunrise */ 3:
                    message.sunrise = reader.float();
                    break;
                case /* float sunset */ 4:
                    message.sunset = reader.float();
                    break;
                case /* float time */ 5:
                    message.time = reader.float();
                    break;
                default:
                    const u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(
                            `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
                        );
                    const d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(
                            this.typeName,
                            message,
                            fieldNo,
                            wireType,
                            d
                        );
            }
        }
        return message;
    }
    internalBinaryWrite(
        message: AppTime,
        writer: IBinaryWriter,
        options: BinaryWriteOptions
    ): IBinaryWriter {
        /* float dayLengthMinutes = 1; */
        if (message.dayLengthMinutes !== 0)
            writer.tag(1, WireType.Bit32).float(message.dayLengthMinutes);
        /* float timeScale = 2; */
        if (message.timeScale !== 0)
            writer.tag(2, WireType.Bit32).float(message.timeScale);
        /* float sunrise = 3; */
        if (message.sunrise !== 0)
            writer.tag(3, WireType.Bit32).float(message.sunrise);
        /* float sunset = 4; */
        if (message.sunset !== 0)
            writer.tag(4, WireType.Bit32).float(message.sunset);
        /* float time = 5; */
        if (message.time !== 0)
            writer.tag(5, WireType.Bit32).float(message.time);
        const u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(
                this.typeName,
                message,
                writer
            );
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message rustplus.AppTime
 */
export const AppTime = new AppTime$Type();